+++ code/Triangle/Model/src/main/java/ru/unn/agile/Triangle/Model/Triangle.java
