+++ code/Triangle/Model/src/test/java/ru/unn/agile/Triangle/Model/TriangleTest.java
