+++ code/Triangle/View/src/main/resources/ru/unn/agile/Triangle/View/Calculator.fxml
