+++ code/Triangle/View/src/test/java/ru/unn/agile/Triangle/View/FakeViewTests.java
@@ -1,0 +1,13 @@
+package ru.unn.agile.Triangle.View;
+
+import org.junit.Ignore;
+import org.junit.Test;
+
+import static org.junit.Assert.fail;
+
+public class FakeViewTests {
+    @Ignore @Test
+    public void fakeTest() {
+        fail("No tests for view");
+    }
+}
