--- code/triangle/View/src/main/java/ru/unn/agile/triangle/view/Main.java
@@ -1,22 +1,0 @@
-package ru.unn.agile.Triangle.view;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class Main extends Application {
-
-    @Override
-    public void start(final Stage primaryStage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("Calculator.fxml"));
-        primaryStage.setTitle("Triangle");
-        primaryStage.setScene(new Scene(root));
-        primaryStage.show();
-    }
-
-    public static void main(final String[] args) {
-        launch(args);
-    }
-}
